Rename exporters/trace/zipkin
