+++ .github/dependabot.yml
@@ -168,7 +168,7 @@ updates:
       interval: weekly
   -
     package-ecosystem: gomod
-    directory: /exporters/trace/zipkin
+    directory: /exporters/zipkin
     labels:
       - dependencies
       - go
