+++ example/jaeger/go.mod
@@ -38,7 +38,7 @@ replace go.opentelemetry.io/otel/exporters/otlp => ../../exporters/otlp
 
 replace go.opentelemetry.io/otel/exporters/stdout => ../../exporters/stdout
 
-replace go.opentelemetry.io/otel/exporters/trace/zipkin => ../../exporters/trace/zipkin
+replace go.opentelemetry.io/otel/exporters/zipkin => ../../exporters/zipkin
 
 replace go.opentelemetry.io/otel/internal/tools => ../../internal/tools
