+++ exporters/zipkin/doc.go
@@ -17,4 +17,4 @@
 // This package is currently in a pre-GA phase. Backwards incompatible changes
 // may be introduced in subsequent minor version releases as we work to track
 // the evolving OpenTelemetry specification and user feedback.
-package zipkin // import "go.opentelemetry.io/otel/exporters/trace/zipkin"
+package zipkin // import "go.opentelemetry.io/otel/exporters/zipkin"
