+++ exporters/zipkin/zipkin.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package zipkin // import "go.opentelemetry.io/otel/exporters/trace/zipkin"
+package zipkin // import "go.opentelemetry.io/otel/exporters/zipkin"
 
 import (
 	"bytes"
